refactor: S3등 다른 저장소를 이용할 때를 대비해 미리 인터페이스로 분리
